Add files via upload
The CovidDataExploration_Project1 SQL script is used for data analysis on COVID-19 cases, deaths, and vaccinations by location and date. The data is queried from two tables: CovidDeathsql and CovidVaccinationsql which are part of the PortfolioProject database. Below is a step-by-step breakdown of what the code does:

The first step is to select data from the CovidDeathsql table where the continent column is not empty. The columns selected are Location, date, total_cases, new_cases, total_deaths, and population.

The script then alters the data type of certain columns in both CovidDeathsql and CovidVaccinationsql tables to ensure proper data handling in the subsequent queries. For example, the total_cases, total_deaths, population, new_cases, and new_deaths columns in CovidDeathsql are all altered to integer or bigint. Similarly, the new_vaccinations and date columns in CovidVaccinationsql are altered to integer and date, respectively.

The script then executes a series of analytical queries to provide insights into the COVID-19 data:

Total cases vs Total deaths: This query calculates the percentage of deaths per total cases.
Total cases vs Population: This query calculates the infection density, which is the percentage of the total cases per population.
Countries with highest infection rates compared to population: This query finds the countries with the highest infection density.
Countries with the highest death count: This query determines the locations with the maximum number of total deaths.
Continent with the highest death count per population: This query determines the continent with the maximum number of total deaths.
Global numbers: This query calculates total cases, total deaths, and the death percentage globally.
Total Population vs Vaccinations: This query calculates the percentage of the population that has been vaccinated.
The script then uses the sp_help stored procedure to get the details of the CovidDeathsql table.

Finally, the script creates a series of views for each of the analytical queries executed earlier. These views make it easier to query and visualize the data in external tools like Tableau
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

The Housing Project script conducts an extensive data transformation and cleansing process on a housing data set. The data set is stored in a table named 'Housing' in the 'PortfolioProject' database. The transformation process includes the following tasks:

Modifying the data types of several columns like 'SaleDate', 'UniqueID', 'SalePrice', 'Acreage', 'LandValue', 'BuildingValue', 'TotalValue', 'YearBuilt', 'Bedrooms', 'FullBath', and 'HalfBath'.
Populating 'PropertyAddress' data for rows where this field is null by leveraging existing data from other rows with the same 'ParcelID'.
Breaking down the 'PropertyAddress' and 'OwnerAddress' fields into separate columns for 'Address', 'City', and 'State' for more granular analysis.
Converting the 'SoldAsVacant' field from 'Y' and 'N' to 'Yes' and 'No' for better readability.
Removing duplicate rows based on a group of columns ('ParcelID', 'PropertyAddress', 'SalePrice', 'SaleDate', 'LegalReference').
Dropping unused columns, specifically 'OwnerAddress', 'TaxDistrict', 'PropertyAddress', and 'SaleDate'.
After the data cleansing, all the columns that were previously altered to different data types are again altered back to 'nvarchar(255)'.