Add files via upload
Calculating The win Percentage for a team (Manchester United): This section of code is calculating the win, loss, draw, total matches and win percentage for the football team 'Manchester United' by season. The 'Wins', 'Losses' and 'Draws' are counted by comparing the team names and scores from the 'Matches' table. Finally, the season, number of wins, draws, losses, total matches, and win percentage are returned for each season, sorted by season.

Average spi per season for a team (Manchester United): In this section, the code is calculating the average SPI (Soccer Power Index, a measure of team's strength) for 'Manchester United' per season. It calculates this by averaging the SPI values from the 'Matches' table where 'Manchester United' is either the first or the second team. The result is the average SPI per season, sorted by season.

Average Spi Per Team in The South African ABSA Premier League in 2022: This section calculates the average SPI for each team in the South African ABSA Premier League for the 2022 season. The SPIs for each team are averaged and then the results are sorted in descending order by the average SPI.

Highest and lowest performing teams per season: In this final section, the code is determining the highest and lowest performing teams per season based on whether they met, underperformed or overperformed their expected goals (xg). The 'PerformanceCounts' subquery counts the number of each performance type (underperformed, overperformed, met expectation) for each team in each season and league. The 'PerformanceRanks' subquery then ranks these counts within each season, league, and performance type. The main query then returns the team with the highest count (i.e., rank = 1) for each combination of season, league, and performance type. The result is a list of the highest and lowest performing teams per season, sorted by season, league, and performance.